+++ Dev/DataModels/Atdi.DataModels.Sdrn.CalcServer.Internal/Iterations/RefSpectrumByDriveTestsCalc/CalcRefSpectrum/Output/ResultRefSpectrumByDriveTests.cs
@@ -15,8 +15,8 @@ namespace Atdi.DataModels.Sdrn.CalcServer.Internal.Iterations
         public long IdSensor { get; set; }
         public string GlobalCID { get; set; }
         public double Freq_MHz { get; set; }
-        public float Level_dBm { get; set; }
-        public float Percent { get; set; }
+        public double Level_dBm { get; set; }
+        public double Percent { get; set; }
         public DateTimeOffset DateMeas { get; set; }
     }
 }
